refactor(core): change TimeWindow.duration_ms from int to int64_t for type consistency
This change addresses type inconsistency in the TimeWindow_T struct where
duration_ms was declared as int (32-bit) while window_start_ms uses int64_t
(64-bit), creating potential issues in arithmetic operations.

Changes:
- Update TimeWindow_T.duration_ms field from int to int64_t
- Update TimeWindow_init() parameter type from int to int64_t
- Update timewindow_clamp_to_duration() to accept int64_t duration
- Update local variables in implementation to use int64_t
- Fix format strings in SocketHTTP2-stream.c to use PRId64 for int64_t

Benefits:
- Eliminates type mismatches in arithmetic operations
- Removes need for explicit casts between types
- Supports window durations longer than ~24 days (INT_MAX ms)
- Consistent with time-related field types throughout the module

Closes #620

+++ include/core/TimeWindow.h
@@ -57,7 +57,7 @@ typedef struct TimeWindow_T
   int64_t window_start_ms;   /**< Start time of current window */
   uint32_t current_count;    /**< Events in current window */
   uint32_t previous_count;   /**< Events in previous window */
-  int duration_ms;           /**< Window duration in milliseconds */
+  int64_t duration_ms;       /**< Window duration in milliseconds */
 } TimeWindow_T;
 
 /**
@@ -66,7 +66,7 @@ typedef struct TimeWindow_T
  * Sets up the window with zero counts starting at the given time.
  * Duration is clamped to minimum of 1ms if invalid.
  */
-extern void TimeWindow_init (TimeWindow_T *tw, int duration_ms, int64_t now_ms);
+extern void TimeWindow_init (TimeWindow_T *tw, int64_t duration_ms, int64_t now_ms);
 
 /**
  * @brief Record an event in the time window
